fix(primitives): remove virtual, move useVirtualizerModal to ts-table
