+++ packages/docs/pages/components/index.mdx
@@ -50,7 +50,6 @@
   <RelatedComponent title="Textarea" href="/components/textarea/api-reference"/>
   <RelatedComponent title="ToastStack" href="/components/toast-stack/api-reference"/>
   <RelatedComponent title="Tooltip" href="/components/tooltip/api-reference"/>
-  <RelatedComponent title="Virtual" href="/components/virtual/api-reference"/>
   <RelatedComponent title="VisuallyHidden" href="/components/visually-hidden/api-reference"/>
   <RelatedComponent title="Calendar" href="/components/calendar/api-reference"/>
   <RelatedComponent title="DateField" href="/components/date-field/api-reference"/>
