--- packages/docs/pages/components/virtual.mdx
@@ -1,5 +1,0 @@
-# Virtual
-
-Augmented forwardRef function to support generics.
-
-![Vritual example](public/assets/overview-virtual.webp)
