--- packages/docs/pages/components/virtual/_meta.json
@@ -1,5 +1,0 @@
-{
-  "api-reference": "API Reference",
-  "virtual-container": "VirtualContainer",
-  "virtual-item": "VirtualItem"
-}
