--- packages/docs/pages/components/virtual/api-reference.mdx
@@ -1,33 +1,0 @@
-# API Reference
-
-Renders a virtual container
-
-```jsx
-<Virtual count={100} estimateSize={50} overscan={5}>
-  {({ index }) => <div>{index}</div>}
-</Virtual>
-```
-
-## Parameters
-
-### `props`
-
-- Type: `@vtex/shoreline-primitives.VirtualProps & @types/react.React.RefAttributes<typescript.HTMLDivElement>`
-
-## Props
-
-### `count`
-
-- Type: `number`
-
-### `dynamic` (optional)
-
-- Type: `boolean`
-
-### `estimateSize` (optional)
-
-- Type: `{ (index: number): number; }`
-
-### `overscan` (optional)
-
-- Type: `number`
