--- packages/docs/pages/components/virtual/virtual-container.mdx
@@ -1,5 +1,0 @@
-# VirtualContainer
-
-Augmented forwardRef function to support generics.
-
-![No-visual component](public/assets/no-visual-component.webp)
