--- packages/docs/pages/components/virtual/virtual-container/_meta.json
@@ -1,1 +1,0 @@
-{ "api-reference": "API Reference" }
