--- packages/docs/pages/components/virtual/virtual-container/api-reference.mdx
@@ -1,29 +1,0 @@
-# API Reference
-
-Renders a virtual container
-
-```jsx
-<Virtual count={100} estimateSize={50} overscan={5}>
-  {({ index }) => <div>{index}</div>}
-</Virtual>
-```
-
-## Parameters
-
-### `props`
-
-- Type: `@vtex/shoreline-primitives.VirtualContainerProps & @types/react.React.RefAttributes<typescript.HTMLDivElement>`
-
-## Props
-
-### `asChild` (optional)
-
-- Type: `boolean`
-
-### `children`
-
-- Type: `@types/react.React.ReactNode`
-
-### `virtualizer`
-
-- Type: `@vtex/shoreline-primitives.UseVirtualizerModelReturn`
