--- packages/docs/pages/components/virtual/virtual-item.mdx
@@ -1,5 +1,0 @@
-# VirtualItem
-
-Augmented forwardRef function to support generics.
-
-![No-visual component](public/assets/no-visual-component.webp)
