--- packages/docs/pages/components/virtual/virtual-item/_meta.json
@@ -1,1 +1,0 @@
-{ "api-reference": "API Reference" }
