--- packages/docs/pages/components/virtual/virtual-item/api-reference.mdx
@@ -1,15 +1,0 @@
-# API Reference
-
-Renders a virtual item
-
-```jsx
-<Virtual count={100} estimateSize={50} overscan={5}>
-  {({ index }) => <VirtualItem>{index}</VirtualItem>}
-</Virtual>
-```
-
-## Parameters
-
-### `props`
-
-- Type: `@vtex/shoreline-primitives.VirtualItemsProps & @types/react.React.RefAttributes<any>`
