--- packages/primitives/src/virtual/index.ts
@@ -1,4 +1,0 @@
-export * from './virtual'
-export * from './virtual-item'
-export * from './virtual-container'
-export * from './use-virtualizer-model'
