+++ packages/theme-sunrise/src/components/index.scss
@@ -37,7 +37,6 @@
 @import 'text.scss';
 @import 'textarea.scss';
 @import 'tooltip-popover.scss';
-@import 'virtual.scss';
 @import 'empty-state.scss';
 
 @import 'calendar/index.scss';
