--- packages/theme-sunrise/src/components/virtual.scss
@@ -1,27 +1,0 @@
-[data-sl-virtual-item][data-sl-virtual-dynamic='false'] {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: var(--sl-virtual-item-size, 0rem);
-  transform: translateY(var(--sl-virtual-item-start, 0px));
-}
-
-[data-sl-virtual-container][data-sl-virtual-dynamic='true'] {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  transform: translateY(var(--sl-virtual-item-start, 0px));
-}
-
-[data-sl-virtual] {
-  height: 31.25rem;
-  overflow-y: auto;
-  width: 100%;
-}
-
-[data-sl-virtual-compose] {
-  position: relative;
-  height: var(--sl-virtual-total-size);
-}
