+++ packages/ts-table/src/index.ts
@@ -1,3 +1,8 @@
 export { getExpandedColumn, getSelectionColumn } from './columns'
 export { TsTable } from './ts-table'
 export type { TsTableProps } from './ts-table'
+export { useVirtualizerModel } from './use-virtualizer-model'
+export type {
+  UseVirtualizerModelProps,
+  UseVirtualizerModelReturn,
+} from './use-virtualizer-model'
